requirements.txt
aiofile
aiogram[i18n]
asyncpg
cachetools
fastapi
httpx
psycopg2-binary
pyyaml
requests
SQLAlchemy
uvicorn
